refactor: 💡 convert Node .js to .ts to fix Jest transpilation
This is a temporary "fix", it converts .js file to .ts, because now Jest
stopped to transpile this file, which caused tests on Node 6 and 7 to
fail because they don't support ES6 features in this file. In general,
all error handling should be revisited, because Node has moved ahead
with what errors it throws.